% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{line_plot}
\alias{line_plot}
\title{Function for plotting Two Dimension}
\usage{
line_plot(x, y, point = TRUE)
}
\arguments{
\item{x}{A numeric, represents the x-axis}

\item{y}{A numeric, represents the y-axis}

\item{point}{A bolean, if set to TRUE (default), add geom_point to the plot}
}
\value{
A ggplot2 line chart
}
\description{
The function create a two dimension line plot with ggplot2
}
\details{
The function get two inputs, x and y, and returns a line plot
}
\examples{

line_plot(x = c(1:100), y = rnorm(n = 100, mean = 10, sd = 2))
# Without geom_point
line_plot(x = c(1:100), y = rnorm(n = 100, mean = 10, sd = 2), point = FALSE)

}
